pyproject.toml
[project]
name = "krishimitra-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clerk-backend-api>=3.3.1",
    "fastapi>=0.118.0",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.8",
    "pydantic>=2.11.9",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "supabase>=2.20.0",
    "uvicorn>=0.37.0",
]
